+++ changelog.md
@@ -8,6 +8,7 @@
 * Add new 'throw_validation_exceptions_on_dispatch' config option
 * Add new 'translate_fail_handler' config option and default callable to dispatch a 'pubsub.events.translation_failure' event
 * Add new 'listen_expr_fail_handler' config option and default callable to dispatch a 'pubsub.events.listen_expr_failure' event
+* Add new 'validation_fail_handler' config option and default callable to dispatch a 'pubsub.events.validation_failure' event
 
 ## 2.0.2 - 2017-05-16